Add use_analysis_sst namelist option (#20)
This option is needed in order to ensure that nudged runs are using the surface temperature from the GFS analysis. It works by setting the physics surface temperature equal to the dynamical core surface temperature over ocean points. The dynamical core's surface temperature is already set to the analysis surface temperature when nudging is active.