remove configure.fv3 from git repo (#21)
`configure.fv3` is a build artifact which results from copying the desired configuration file to that location. Right now, it's set to an out of data file which breaks the build. This PR deletes the file and adds it to .gitignore. This changes nothing for this repo, but prevents a bug I ran into in `fv3gfs-python`.