reduce number of run steps in dockerfile (#11)
A copy of every layer is stored on your system when building dockerfiles, if caching is enabled. We use caching since we re-build these images a lot. Reducing the number of run steps reduces the number of copies of the image stored on our system, which in turn reduces the number of times we need to run `docker system prune`.

I've combined any RUN operations that could be combined in the docker file. No other changes.

Commit history:

* reduce number of run steps in dockerfile

* use absolute path for copy target to avoid error

* fix data deletion directory before artifact upload

* update circleci references with new output from last PR