[VCMML-469] Add option to add precipitation implied by nudging to surface precipitation (#53)
This PR makes it standard that minus the column integrated moistening implied by nudging specific humidity to NCEP analysis is added to the surface precipitation felt by the land surface model.  If nudging is not active (or if nudging is active and specific humidity is not nudged) nothing changes.

It's largely a translation of https://github.com/VulcanClimateModeling/fv3net/pull/509 into fortran.  Like in https://github.com/VulcanClimateModeling/fv3net/pull/509 we:
- Make sure to convert the column integrated moistening from nudging to the appropriate units
- Make sure to clip the updated precipitation at zero, so that the land surface does not feel negative precipitation
- Apply the change at the end of the physics.