[VCMML-487] Add install makefile target (#54)
The build system of the python-wrapper is somewhat tightly bound to this repository. One issue is that the python wrapper hardcodes paths to specific `.o` and `.a` files produced by this repo and to the NCEPlibs. This PR helps decouple this aspect of the build by installing the libraries in standard location and by adding some metadata to track all the linking flags needed by consuming packages.

Changes:
- Add an `install` target for non-serialbox build. This allows installing the build artifacts in a system location for easier compiling with fv3gfs-python
- Creates a pkg-config configuration for the fv3gfs libraries that can then be called from the `setup.py` of the python wrapper.

Related https://github.com/VulcanClimateModeling/fv3gfs-python/pull/117