[VCMML-679] Correct "omega" diagnostic to be mass-conserving in the non-hydrostatic model (#102)
This PR adds a new mass-conserving diagnostic for the pressure velocity, which is computed based on the outline here (the original does not conserve mass in the non-hydrostatic version of the model). Essentially this computes omega in non-hydrostatic mode the same way it is computed in hydrostatic mode. We replace the original "omega" diagnostic with this new calculation and rename the old omega diagnostic "approximate_omega."

This diagnostic tends to behave more reasonably in the sense that it averages out more closely to zero in the global mean where the model uses pure pressure vertical levels.