Refactor diag_manager-controlled diagnostics for backwards compatibility (#140)
This PR refactors the diagnostics-manager-controlled diagnostics introduced in #112 such that they are defined within the same data structure as the other physics diagnostics.  It does so by simply adding a flag to the derived data type (default `.false.`) denoting whether a particular diagnostic is `diag_manager_controlled` or not.  The code path taken for sending data to the diagnostics manager then depends on the value of this flag.  It should not change any answers or functionality (regression tests confirm this).

xref: https://github.com/VulcanClimateModeling/fv3gfs-wrapper/issues/230