Remove bespoke logic in gfsphysics makefile
The gfsphysics makefile had some debugging related logic, to allow
viewing ./GFS_layer/GFS_physics_driver.F90 etc after pre-processing.
This logic meant that CPP needs to be configured, and made the model
harder to build in new contexts (see e.g. #46). For example, using the
wrong `cpp` can result in a failure if it doesn't set the __GFORTRAN__
macro.

I don't think we are debugging this part of the model, so don't need
this bespoke code. The model compiles fine without it (the .o files are
covered by the default makefile pattern rules).