Add variables we can use to modify the radiative fluxes to the Statein (#158)
This PR adds three variables to the `Statein` derived type that we can use to override the downward longwave, downward shortwave, and net shortwave radiative fluxes at the surface.  By default these variables are not allocated or used.  To use them one must set the `gfs_physics_nml.override_surface_radiative_fluxes` namelist parameter to `.true.`.   Setters for these variables in the Python wrapper can be found in a [complementary PR](https://github.com/VulcanClimateModeling/fv3gfs-wrapper/pull/244).

A significant amount of work in this PR also goes into refactoring the diagnostics related to these variables.  When running in a mode where we override these fluxes, we want the usual fortran diagnostics  -- namely `DLWRF`, `DSWRF`, `USWRF`, `DLWRFI`, `DSWRFI`, and `USWRFI` -- to correspond to the overriding fluxes.  We modify things such that this is the case.  In addition, to enable diagnosis of what the traditional radiation code would predict for these fluxes, we add diagnostics with the suffix `_from_rrtmg`.  [This notebook](https://github.com/VulcanClimateModeling/explore/blob/master/spencerc/2021-03-04-predicting-radiative-fluxes-online/2021-03-11-verify-diagnostics.ipynb) checks to make sure all of these diagnostics work as expected in a case where we override the fluxes with ML (the regression tests ensure that things continue to work when we are not overriding the fluxes).

Care has been made to ensure that if `gfs_physics_nml.override_surface_radiative_fluxes` is set to `.false.`, the model will run identically to before, both in producing bit-for-bit identical results and in terms of memory allocated during runtime.  When it is set to `.true.` the model will use more memory to accommodate the arrays used for storing the RRTMG diagnostics as well as the overriding fluxes in the `Statein`.