nix: build fv3 with nix (#170)
Nix is a package manager with a unique idea. It's builds are guaranteed to be deterministic and packages are described by functions in a DSL. This provides a level of composability that is awkward with multistage docker files. 

To test this functionality, start with a clean clone of fv3gfs-fortran and follow the instructions in the README.

Some cool things:
- nix packages can be composed and modularized across several file.
  - We only need to host ~200 mb of binaries for FV3, ESMF, and FMS vs multiple GB docker image
- bare metal 🤘 🎸 
- Reliable caching. The same  caches can be used locally and in CI. Nix packages always resolve to a single hash (see all the hashes in the .nix files), so the caching almost always works.
- available on mac 🍎  and linux 🐧 
- easier to run than a container (no need for bind mounts permissions etc)
- completely deterministic builds. If the binaries stop being hosted, we can still build from source with the exact same command `nix-build`.

Some not so great things
- Nix is a little weird/less widely known than docker. I personally don't think nix is more complex than e.g. debian, anaconda or spack packages.

I am happy to fix bugs, and this repo moves pretty slowly/has static dependencies. While we could modify the docker builds to use this approach (or even use nix to build docker images), I am not proposing we do that right now.  I just want to see this code in the wild and use these packages personally. I know some others on the team e.g. @jdahm have played around with nix, so it'll be interesting to see if it catches on.