Fix coarse geopotential height diagnostic (#160)
The coarse geopotential height was computed incorrectly; among other things, it did not take into account the surface height.  This fixes things by using the `get_height_field` subroutine from `fv_diagnostics.F90` to compute the height.  This means that we are coarse-graining the exact field that is output as a fine-grid diagnostic.  I've verified this [here](https://github.com/VulcanClimateModeling/explore/blob/master/spencerc/2021-02-24-fix-coarse-z/2021-02-24-fix-coarse-z.ipynb).

As expected, this requires an update to the checksums for the `model-level-coarse-graining` and `pressure-level-coarse-graining` tests, since those both output a coarse geopotential height diagnostic.