Add diagnostics of column-integrated nudging of t, u, v (#166)
It is useful to be able to output the column-integrated nudging tendencies, e.g. so they can be saved at a higher frequency than the 3D nudging tendencies. This is already implemented for nudging of humidity. This PR adds column integrated nudging tendencies of temperature, eastward wind and northward wind.