use git for fv3net source (#189)
This means that unstaged changes will not incur any rebuilds, and files
that aren't in version control (e.g. build artifacts) will not cause nix
to rebuild.

If you do want something to be included in a rebuild, it will need to be
staged to the git index (e.g. using git add).