emulation: save state after gscond
The temperature and humidity states after gscond are used to compute a relative
humidity tendency in the following timestep. Also, we always run the model with
the same physics (dtp) and dynamics (dtf) timestep, so the tp1 and cp1 variables
are identical to cp. dtp is actually set equal to dtf in the model startup.
See this
[notebook](https://github.com/ai2cm/explore/blob/master/noahb/emulation/microphysics/2021-11-03-verify%20tbd%20data.ipynb)
for proof.

* Remove the redundnant call py fort state setting for tp1_cpf, etc
* Rename the output of `tp` etc to be more accurate
* Save the temperature and humidity state after `gscond`
* Install black in nix environment to make development easier