Add option to derive net shortwave radiative flux in physics driver
When using ML to predict the surface radiative fluxes, it is important to use
an albedo that is at least consistent with the coarse model when computing the
net shortwave radiative flux at the surface.  This adds an option to do this by
computing the absorption (i.e. 1 - albedo) using the RRTMG fluxes, and makes
the assumption that the absorption is zero if the RRTMG downward shortwave flux
at the surface is zero.