Streamline makefiles
The makefiles here were pretty convoluted with recursive behavior
(including the FV3/makefile calling itself). There is also overhead
associated with passing build flags to the wrapper's makefiles. This
commit flattens this hierarchy.

* remove self-recursive trigger of "all"

This was enabled by make `libs` a dependency of all FV3/*.o.
Since libs is a phony target, it will be built once per invocation of make,
which will retrigger the builds of the FV3/.o files. This doesn't seem like a
big deal since most development happens upstream of this point anyway (e.g. in
the gfsphysics subdirectory) and will reinvoke the build regardless. The FV3/.o
files compile in a few seconds anyway.

* remove recursive makefile logic in wrapper builds
* remove duplicated logic for atmos_model.o etc
this is already handled by mkDepends.pl and the existing pattern rules in
conf/make.rules