Fix zhao_carr_physics diagnostics
This diagnostics has a bug when emulate_gscond_only is true.  In this
case, zhao_carr_physics is identical to zhao_carr_emulator. This
happens, because Diag%zhao_carr_physics should be defined as sum of the
precipitation and condensation updats from the fortran. Currently, it is
the sum of the emulated condensation and the physics precip.

* Fix this bug