Make indefinite output the default behavior
This also changes the name of the parameter and its sign.  If outputting
physics diagnostics indefinitely is now the default behavior, it feels
more intuitive if one needs to opt-in to using the fdiag behavior with the rest
of its associated parameters, rather than needs to opt-out of outputting
physics diagnostics indefinitely.