Set the model to output physics diagnostics indefinitely by default (#302)
Originally, by default, if `fdiag` was provided as a scalar value and `fhouthf`  was less than zero, the physics diagnostics would be output on regular intervals of length `fhout` hours for a finite number of total simulated hours, governed by `fhmax`.  This can be problematic in long simulations.  Simply using a large default value of `fhmax` is not tenable, because we cannot allocate an infinitely large `fdiag` array (which is what the original physics diagnostics approach uses under the hood).  

This PR changes the default behavior such that physics diagnostics will now be output indefinitely on a regular interval of `fhout` hours, ignoring any values set for the `fdiag`, `fhmax`, `fhmaxhf`, or `fhouthf` parameters.

While in practice we do not use the ability of `fdiag` to output physics diagnostics on irregular intervals, if one would like to restore this behavior, one can set the `atmos_model_nml.use_fdiag` namelist flag added in this PR to `.true.`.  Naturally the model will then use the parameters specified for `fdiag`, `fhmax`, `fhmaxhf`, and `fhouthf` in conjunction with `fhout` to determine when and how long to output physics diagnostics.

Resolves #301.