Port SST prescribing code from SHiELD (#304)
This PR ports the code to prescribe a time-varying SST (and optionally sea ice) pattern from a file from the public [SHiELD_physics repository](https://github.com/NOAA-GFDL/SHiELD_physics).  It also adds a regression test.  The regression test is fairly involved, but doubles as documentation for how to use this feature of the model which has a few gotchas that are described in the testing code.