Update gscond-only case to truly be gscond-only
Previously, we were in the scenario where gscond-only was true even
when emulating both schemes.  This led to some confusing behavior in
the diagnostics where when running offline, we were still removing
the offline emulation from the Stateout to determine the full zc
tendency update, which is incorrect.  For true gscond-only, we want
our online fortran precpd to respond to the gscond output, there is
no piggy-back analog because there is no emulation run for precpd.
Therefore, I removed the adjustment code from the total physics.
This does not fix the gscond+precpd emulation yet but makes gscond-only
consistent.