Update full zc emulation to correct diagnostics
Now when running with emulate_zc_microphysics=True we expect gscond_only
to be false.  The state is no longer updated until after the fortran
gscond and precpd have completed so those diagnostics will be correct.
We still calculate the gscond emulation diags for offline/online case
and the the final total emulation tendency.