Add column integrated physics tendency diagnostics (#363)
This PR adds column-integrated physics tendency diagnostics for temperature and specific humidity.  In the case of temperature we follow what was done for the column integrated nudging and fv_sat_adj tendencies, where we scale the column integrated result by either Cp or Cv depending on whether the model is run hydrostatically or non-hydrostatically.  As in those cases we do not take into account moist effects when computing these values.  That's something we could eventually think about modifying, but might be better for a future PR.

Note: I moved the `compute_column_integral` subroutine to the `fv_diagnostics` module to work around a circular import.

[This notebook](https://github.com/ai2cm/explore/blob/master/spencerc/2023-04-07-column-physics-diagnostics/2023-04-07-test-column-physics-diagnostics.ipynb) verifies that these diagnostics match what we would get if we computed them offline.