Bump base Ubuntu image
This bumps the base Ubuntu image used to one with a version of 20.04. The
pinned SHA makes things consistent with the base image we use when building the
prognostic run in fv3net.

Note that due to NOAA-EMC/fv3atm#33, the tests with the TKE-EDMF scheme active
fail in debug mode.  We currently mark them as expected failures.  To
demonstrate that answers do not change when we fix this bug, we run the
coarse-graining tests in repro mode here and checkpoint the checksums.  In a
followup commit, we will fix the bug, unmark the relevant debug mode tests, but
retain the repro mode tests to demonstrate that the tests still all pass
without modifying those checksums.  Finally in one last commit we will remove
those temporary repro mode tests as well as their checksums, as they will no
longer be needed.

Because the serialize_image fails to build with the new compilers, we also
elect to remove the infrastructure and tests associated with it.