Add surface type fraction diagnostics (#377)
This PR adds diagnostics-manager-controlled surface-type-fraction diagnostics for the `ocean_fraction`, `land_fraction`, and `sea_ice_fraction`.  These are all added as part of the `gfs_sfc` diagnostics module, similar to the `slmsksfc` diagnostic containing the integer-encoded surface type.  As stated in #371, the advantage of these diagnostics is that they can be sensible time-and-area-averaged, so they retain their value upon standard time-averaging, coarse-graining, and conservative regridding approaches.  

[This notebook](https://github.com/ai2cm/explore/blob/master/spencerc/2023-08-21-surface-type-fraction/2023-08-21-test-surface-type-fraction-diagnostics.ipynb) shows that these diagnostics are working as expected.

Note that it is theoretically possible to run FV3GFS with fractional land (this PR assumes one is always running in the mode of one grid cell, one surface type).  Since we have not explored running in that mode yet, we have elected to omit trying to handle that case in this PR (it would make things more complicated).

Resolves #371.